Fix a bug in URI generation for Resources
The 'eDirectResource' method of SmartObjects always returns its eResource, even when it is not  'directly' contained by said eResource.
This leads to the assumption that the SmartObject is a root element.
As a result, this method returns '/' for practically all SmartObjects because the getURIFragmentRootSegment() method returns an empty string for non-root elements.